Refactored Item and File objects into their own modules.
